requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==1.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic-bin==0.4.14
pypdf==3.17.0
python-docx==1.1.0
yt-dlp==2023.11.16
pydub==0.25.1
openai==1.3.3
elevenlabs==0.2.21
requests==2.31.0
pydantic==2.5.0